Revert "Make easy easier."
This reverts commit 4cceccc70c9f9a845ac10602b6e3874236c72ff1.

+++ Kong/src/entities/Clappers.h
@@ -13,15 +13,15 @@ class Clappers {
         Clapper & getClapper(uint8_t index);
 
         void updatePositions();
-        void setDelayMax(uint16_t delayMax, bool updateDelay);
+        void setDelayMax(uint8_t delayMax, bool updateDelay);
         void launchClapper();
         void reset(uint8_t y);
 
     protected:
 
         Clapper clappers[4];
-        uint16_t delay;
-        uint16_t delayMax;
+        uint8_t delay;
+        uint8_t delayMax;
 
 };
